+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `EncCBOR` and `DecCBOR` instances for `ScriptContext` for PlutusV4
 * Change `DecCBOR` instance for `LeiosCert` to reject a signers bitfield larger than `maxLeiosCertSignersBytes`
 * Widen `cardano-crypto-class` upper bound to `<2.7`
+* Add `EncCBOR` and `DecCBOR` instances for `LeiosSeat` and `LeiosCommittee`
 
 ### `testlib`